util/backtrace: introduce make_backtraced_exception_ptr()
This is a similar to the existing `throw_with_backtrace()` but is geared
towards being used with `make_exception_future<>()`. As such it returns
an `std::exception_ptr`. As `make_backtraced_exception()` is basically a
subset of what `throw_with_backtrace()` does, the latter is
reimplemented in terms of the former.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20200408075222.700717-1-bdenes@scylladb.com>