timer: add scheduling_group awareness
Timers currently always execute in the default scheduling group.
This means that if they launch any continuations, these continuations
will also execute in the default scheduling group, regardless of the
scheduling group that the timer's creator belonged to. This can cause
a "leak" of scheduling information, causing an fiber with timers
to consume more resources than it is entitled to.

To fix this, capture the current scheduling group when the callback
is captured, and also provide a way to set the scheduling group
for a callback explicitly. When we run a timer callback we change
temporarlily to the captured scheduling group. Note that the timer
is still run within the poller context; any accounting will only
happen to continuations launched from the callback.

Note that seastar::sleep() isn't affected by the bug, since it doesn't
launch callbacks directly (instead it launches callbacks indirectly via
promise::set_value()).

A unit test is provided to check that scheduling groups are
propagated correctly.
Message-Id: <20200413103140.10504-1-avi@scylladb.com>