rpc: always shutdown socket when stopping a client
When stopping a client socket::shutdown() needs to be called to abort
send and receive loop, but it is skipped if _error flag is already set.
Set _error flag does not mean though that both loops already aware
that they need to shutdown.

The path calls the shutdown() unconditionally. It is safe to call
socket::shutdown() if the error left an underlying socket disconnected
since shutdown() does not throw in this case.

Refs scylladb/scylla#6060

Message-Id: <20200414080257.GA27043@scylladb.com>