timer: document
Document the timer class. It is placed into a new "timers"
module - I could not find a better place for it.

The private type callback_t in method signatures was expanded
to noncopyable_function<void ()> - it's not reasonable to use
a private type in a signature, and even making it public requires
extra effort to figure out what it is. I left internal uses as is.
Message-Id: <20200416124718.95132-1-avi@scylladb.com>