fair_queue: remove legacy capacity configuration
In the old days of the fair queue all one needed to configure was a
capacity. Different requests would have different weights, which was
passed as a parameter to queue_request(), that weight was used in
the cost function, and that was it.

Newer incarnations of the fair_queue can operate in terms of requests
and sizes. But to ease compatibility with older versions we kept the
capacity code as is.

There is a simpler way to do this: we can get rid of the capacity code
by teaching the fair_queue user (at the moment, the I/O scheduler), to
express more or less the same idea as before but in terms of sizes and
iops.

To do that, we set weight = read_request_base_count, which the I/O queue
uses to represent a request with baseline cost, and the size is set
to read_request_base_count * len. Note that in the legacy model reads
and writes behave the same.

The IOPS capacity of the legacy queue is then read_request_base_count *
capacity, and we should hit it with small requests. The size limit of
the queue is read_request_base_count * capacity * 16kB, and this is what
we expect to hit with large requests.

Notes:
- The weight function for the legacy queue used to be 1 + (len / 16kB),
  which is why we choose 16kB here

- the write / read ratio multipliers were previously initialized to 1
  by 7afc518d0a9e. In that issue we were just trying to make sure that
  an unconfigured queue had anything other than zero in those fields,
  so we just go ahead and initialize them with the base cost.

Tests: unit (debug)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200420153205.3646-1-glauber@scylladb.com>