http server: fix "Date" header format
RFC 7231, Section 7.1.1.1. Defines a specific format, called the "preferred
format" or "IMF-fixdate", which the "Date:" header must use (it also allows
two additional obsolete formats, which are irrelevant to this discussion).

We "almost" got this format in our HTTP server implementation. Almost, but
not quite - we are missing the day of the week. Some clients will be
forgiving and understand our partial format (four decades ago, when RFC 822
was written, clients were supposed to be forgiving) - but some won't,
and fail to parse our Date header. So we need to fix the code to conform
to RFC 7231.

While reviewing this code, I also noticed it will produce wrong results
when not in the C or English locale - the RFC 7231 date format uses English-
language names. We will fix this pre-existing issue in the next patch.

Fixes #739.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20200421142501.24463-1-nyh@scylladb.com>