http server: fix date function on non-English locales
The HTTP server needs to output the "Date:" header in the RFC 7231
format, and we have a special function to do that. Unfortunately,
our existing implementation uses the C library function strftime(),
which only works as expected (i.e., uses English-language day and
month names) if the current locale uses the English language.
But Seastar, being a library, cannot rely on the current locale to be
anything specific (and cannot change it either).

So in this patch, I just open-code the date formatting with fmt.
It's fairly trivial and just requires the short list of (English) day
and month names.

This patch also moves the date formatting function from the header
file to the source file - it is not a performance-sensitive function
and shouldn't have been in the header file.

Fixes #740.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20200421162312.28682-2-nyh@scylladb.com>