rpc: add Doxygen documentation the protocol class
Although we do have a high level documentation in doc/ for rpc and its
features, this is not discoverable, and not even included in the
generated Doxygen documentation, where I usually look for all
documentation. The protocol class itself and its members is very lightly
documented as well, but mostly with non-doxygen comments, meaning it
doesn't even show up in the generated documentation. As a result I
repeatedly found myself resorting to reading the source to find out how
different aspects of rpc work. To help future myself and others in a
similar position, this patch documents the protocol class, includes the
various doc/rpc*.md in the generated documentation and adds links to
them from the documentation of the protocol class, making them
discoverable. This patch also defines an rpc group. As I wasn't exactly
sure what is and isn't part of the public API (no internal namespace is
used), I only added symbols I was sure of. This means I could have
missed some of them, but this is still better then the current situation
where almost none of the rpc symbols turn up in the generated
documentation.

Tests: manually inspection of the generated documentation
Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20200423085811.144953-1-bdenes@scylladb.com>