io_queue: bring capacity back
Unfortunately io_queue's capacity was a public function, that was
used by other projects. But we don't need the fair_queue to expose it
so we can implement it: capacity is only about how the I/O Queue is
configured, and the I/O queue layer knows how the configuration is
supposed to be.

This patch restore the capacity() method based on that.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200423132102.2010-1-glauber@scylladb.com>