io_queue: deprecated request tracking
Those two functions currently have no user in tree, nor do they have any
users in Scylla. However to be safe I am proposing a 2-step deprecation
process for the benefit of other users, although it is unlikely they are
being used.

These role of these functions is to allow a user to extract metrics from
the I/O queue. However ever since we started using prometheus with
labels, they are essentially just a sum of whatever goes into the
individual queues anyway and are not used nor useful.

The fair_queue will soon lose the ability entirely of expressing
individual requests, so we deprecate that too.

Deprecating both the fair_queue and the io_queue users is safer,
and the double accounting, although not ideal, allow us to cleanly
do it without seeing deprecation messages everywhere and should be
cheap enough not to matter. It also improves decoupling.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200423144642.6700-1-glauber@scylladb.com>