cmake: Add support for testing with valgrind
Ideally we would test with the memory sanitizer, but is hard for a
couple of reasons:

* It is only in clang
* It requires that all dependencies be built with it

While we want to support clang, I think that building all dependencies
with the memory sanitizer is impractical with the current build
system.

This change instead adds support for valgrind. It is crazy slow and
doesn't support all the system calls we use, but it can already detect
a few problem that our current debug build cannot.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>