fstream: Zero the end of a buffer before writing
We are currently writing uninitialized bytes and then truncating. As
far as I can tell a crash between the two operations will leave
uninitialized data on disk.

I don't know if we could use a slower write for this special
case. This patch assumes that the write + truncate is the best we can
do. The patch just initializes the end of the buffer so that the
contents are predictable.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>