http: make headers case-insensitive
According to rfc7230, HTTP headers are case-insensitive,
which was not the case in Seastar HTTP server.

Note that a more optimal solution would be to keep
lowercase strings in the map itself, but since _headers
is a public field, it raises backward compatibilty issues.
A possible follow-up is to provide getters/setters for
the _headers field and eventually make it private.
Then, header names can be transformed to lowercase
when being added to a map, which will in turn avoid
transforming them to lowercase every time they need
to be compared.

Tests: unit(dev), manual with modified apps/httpd server
Fixes #636

Refs https://github.com/scylladb/scylla/issues/6319

Message-Id: <23ab33d1e9f499c8003b84c5df4cd06d97033142.1588166645.git.sarna@scylladb.com>