circular_buffer_fixed_capacity: Always mask indexes
There were a couple places we were not masking the indexes, which
could cause crashes after a _begin underflow.

The move constructor now behaves a bit more like the std::array
one. In particular, the moved values are destructed when the moved
from circular buffer is destroyed.

The clear function was already correct, this just simplifies it a bit.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>