membarrier: fix madvise(MADV_DONTNEED) failure and crash with --lock-memory
On old kernels, we use madvise(MADV_DONTNEED) as a way to generate
inter-processor-interrupts to other cores, forcing them to inject a
memory barrier. However, if the user used --lock-memory (which calls
mlockall()), the madvise() calls fails. Luckily we asserted it doesn't
so we ended up with an early crash rather than a hanging shard.

The fix is to call munlock() on that memory, to ensure madvise() will
succeed.

Ref https://github.com/scylladb/scylla/issues/6346

Reviewed-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200505160527.22951-1-avi@scylladb.com>