distributed_test: Fix it on slow machines
The test before this patch assumed that sleeping for 100 ms would be
enough to bring up concurrency to the limit it is testing.

In a slow machine (like the valgring VM), more time is needed.

This patch changes the test to wait for the specified concurrency and
then sleep for 100 ms. Another advantage of this is that we can be
exact when checking max_concurrent_observed.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200506214047.18284-1-espindola@scylladb.com>