build: use -fcommon compiler flag for dpdk
gcc 10 switched its defaults to -fno-common [1], which dpdk is not
prepared for. Use -fcommon explicitly for dpdk builds. This is safe
as it was the default for gcc 9 and previous.

I see that dpdk upstream did some work to make it -fno-common
friendly, but it is better to wait for the work to complete and
then rebase on a new dpdk version.

[1] https://gcc.gnu.org/gcc-10/changes.html

Reviewed-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200507100340.769412-1-avi@scylladb.com>