rpc: Assert that sinks are closed
Sending data over a sink creates continuations that are only waited
when it is flushed or closed. Failure to wait for them can cause use
of freed memory. It is better to abort instead.

We could be a bit more lax and assert that _sem is empty, but it seems
better to just require the stream to always be closed.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200508011652.75200-1-espindola@scylladb.com>