future: Fix self move assign of futures
It is invalid to call new(&x) future(std::move(x)).

We also don't want to have explicit checks for the self move assign
corner case in a type that is used everywhere.

The new implementation avoids both the undefined behavior and the
explicit check.

Fixes #741

Should also fix the test failures with fedora 32 and c++20.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>