coroutines: enable for C++20
C++20 coroutines are only slightly different from the coroutines TS
so we can support both.

The differences are:
 - header <coroutine> vs. <experimental/coroutine>
 - namespace std vs. std::experimental

Unlike the existing coroutine TS support, I made C++20 support
automatic, without cmake magic. This relies on C++20 <version>
header and C++20 feature detection macros. To support this,
the demo and unit test are now compiled unconditionally, and simply
do nothing if coroutine support is not present.

Note that gcc 10 requires -fcoroutines to enable coroutine support.
The patch makes no attempt to enable this automatically - if gcc
makes coroutines support opt-in, this patch respects it.
Message-Id: <20200512135155.244854-1-avi@scylladb.com>