future: Move promise_base move constructor out of line
This fixes some of the size regression caused by my recent changes to
self move assignment.

Constructors have more code in them than is apparent, so keeping the
promise_base constructor out of line is a win.

This is a 0.4% text size win (from 37805172 to 37644148 bytes) and a
0.65% tps improvement in

perf_simple_query --duration 16 --smp 1  -m4G

over 500 randomized runs (from 160706 to 161764).

The rest of the difference seems to be from what code is or is not
covered by a always_inline, but I was not able to obtain a win in both
code size and performance by shuffling that around.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200512163214.5021-1-espindola@scylladb.com>