do_with: Merge implementations
This removes the single value implementation and adjusts the general
one to also work with a single value.

This fixes the included test as it was supported by the general
implementation but not the specific one.

We now always move every value twice. The specific implementation
would move only once.

The result of the extra move is that scylla's text section increase by
0.17% (from 38765748 to 38834484) and the tps reported by

$ ./perf_simple_query --duration 16 --smp 1  -m4G

over 500 runs with randomized layout and stack is reduced by 1.13%
(from 161314 to 159510).

Both issues will be fixed in a followup patch in this series.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>