future: introduce future::then_unpack()
With the deprecation and anticipated removal of variadic futures,
passing a multi-value payload is done using a future<std::tuple<...>>.
This makes .then() calls cumbersome since they need to unpack the tuple
themselves.

Introduce a .then_unpack() variant of .then() that unpacks a tuple
payload into the continuation's parameter list.

This is intended to easy migration away from varaidic futures,
especially in the case of when_all_succeed().