coroutines: reduce template bloat
To integrate with the compiler, Seastar supplies an
std::[experimenta::]coroutine_traits specialization that
tells the compiler how to generate code for coroutines that
return a seastar::future.

These templates take as parameters the types of the parameters
to the coroutine function, but these are never used. The result
is that we instantiate many more templates than we need. For
example,

    future<> f(int x) {
        co_return;
    }

    future<> g(long x) {
        co_return;
    }

will instantiate

     coroutine_traits<seastar::future<>, int>
     coroutine_traits<seastar::future<>, long>

This is wasteful, so collapse these instantiations by using
a common base type (which simply doesn't have the function
parameter types).

Tests: unit(release), coroutine_demo(release)
Reviewed-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200519174036.1009285-1-avi@scylladb.com>