sharded: support passing arbitrary shard-dependent parameters to service constructors
The sharded<> template supports passing two types of parameters
to service constructors:
  1. arbitrary constant values - these are copied to all instances
  2. other sharded<> values - these are transformed before passing
        to the constructor; the local instance is extracted and passed

These two types are not flexible enough: sometimes one needs to pass
the result of some function that is dependent on the current shard,
perhaps a specific member of a sharded<> instance, or a allocation
that needs to happen on the local shard.

To support this, a third type parameter is introduced: shard_parameter.
It is a wrapper around a lambda and some parameters. The lambda is
executed on each shard to obtain the shard-local value. The parameters
are also treated the same way, so they can refer to sharded<> services
(or other sharded_parameter:s).

Example:

    sharded<foo> s_foo;
    extern sharded<bar> s_bar;
    s_foo.start(
            // will be passed as-is:
            7,
            // will be evaluated on each shard:
            sharded_parameter([] { return std::make_unique<char[]>(1000); }),
            // local instance will be extracted
            std::ref(s_bar),
            // member of local instance will be extracted:
            sharded_parameter(&bar::some_member, std::ref(s_bar))
    ).get();

TODO: constraints, more documentation, examples
Message-Id: <20200520180946.71146-1-avi@scylladb.com>