coroutines: add support for forwarding returns
So far, we support co_return with the value_type of the returned
future: for future<int> we support "co_return 87;". If one wishes
to return a future<int>, one must write

    co_return co_wait function_returning_future();

This adds the overhead of a co_await - the coroutine suspends
while function_returning_future() is executing, resumes after
the function returns, unpacks the return value from the future
and re-packs it into the promise.

This patch adds an alternative method:

    co_return function_returning_future();

Using this method, the coroutine will return immediately,
leaving the function to continue running. The future returned
by the function is forwarded to the coroutine's return value,
so from an outside observer's point of view nothing has changed.

There is still a difference: with co_return co_await any
variables in the coroutine frame are still valid while
the function is running. So with

   int x = 8;
   co_return co_await f(&x);

the pointer is valid until f()'s future becomes ready, while with

   int x = 8;
   co_return f(&x);

the pointer if only valid until f returns (the first point it blocks).
So care must be used if passing coroutine local variables by reference.
Message-Id: <20200524170038.1733938-1-avi@scylladb.com>