reactor: don't mlock all memory at once
Linux runs the mlockall() syscall on one cpu (as is expected),
so faulting in and locking all of a machine's memory can take
many minutes, during which just one cpu is active. This causes
startup to feel sluggish.

Unfortunately, faulting in parallel and running mlock in parallel
on all shards doesn't help, because this code runs under mmap_sem.

The only thing that helps is deferring the locking to fault time,
using the new (Linux 4.4+) flag MCL_ONFAULT. This tells the kernel
to not to fault in missing pages, but rather lock them when they
are faulted. If the application is doing other work, then that other
work will be intermixed with faulting+locking and we won't see
significant throughput degradation. If all the application is doing
is fauling in memory, we'll be slow (but we already were slow in
that case).

Ref https://github.com/scylladb/scylla/issues/6460.
Message-Id: <20200525164746.1870676-1-avi@scylladb.com>