sharded: unify invoke_on_*() variants
We have many invoke_on() variants. One axis of variance is the
operation: invoke_on(), invoke_on_others(), invoke_on_all().
Another axis is how the function is specified: as a pointer to
member (with further options for functions returning void and
functions returning future<>), std::function<>, and functor.

This patch unifies the variants along the second axis. Instead
of different overloads for pointers-to-members and functors,
use a single overload and rely on std::apply() being able
to handle a pointer-to-member as a callable.

We gain more flexibility this way, previously the functor
overload did not accept any arguments, but as a result of the
unification, it does.

Note that the std::function<> overload remains, because it is
used for type erasure.

Concepts are added to catch type errors early.

There is a degradation in one aspect: previously, invoke_on()
was able to disambiguate pointer-to-members with the same name
but different signatures, and now it cannot. There was two
such calls in http_server (that are adjusted by this patch).
I believe such calls are rare (there are none in Scylla) and that
the simplficiation is worth requiring the user to adjust such
calls.

Tests: unit (dev)

I also tested Scylla in C++20 mode, as it has many more call
sites. No adjustments were needed and all tests passed.
Message-Id: <20200526093724.1971830-1-avi@scylladb.com>