tests: add a perf test to measure the fair_queue performance
This is a useful tool to see the performance of the fair_queue
in isolation of the I/O Scheduler.

There are two tests:
- one in which there is one fair queue per shard and the dispatches
  are local. This should be the fastest case

- the other in which sending requests have to go through a coordinator
  shard.

In both cases, the fair_queue can only admit one request at a time.

The end goal of this is to be able to evaluate upcoming patches that
would move the fair queue to a model where dispatch is always local
but it shares capacity with a shared pool. When we do that, we will
want the decrease in performance compared to the fully local case not
to be too big, and hopefully see an increase in performance compared
to the shared case with full dispatch.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Reviewed-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20200605011448.17465-1-glauber@scylladb.com>