thread: Detemplate thread_wake_task
Before this patch thread_wake_task needed to be templated so that it
could set the value in the corresponding future.

Instead, we can keep the original future/promise relationship. The
reason a promise's _state pointer is normally pointed to a task, is
that the future is about to be destroyed (in future::then for
example). That is not the case with a thread_wake_task: the original
future will stay on the stack.

This requires a small update to the existing logic to handle a promise
with both the _state and _task pointers set. The change is to not
clear _state in promise_base::make_read, to maintain the invariant
that when a promise points to a future, _state points to the state in
the future.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>