net: packet: mark move assignment operator as noexcept
The move assignment operator relies only on the move constructor,
which is noexcept, so th assignment operator can be safely marked
noexcept as well.

This fixes the build on gcc 9, where move-assigning a dns_resolver
tcp_entry fails, likely due to a gcc 9 bug (see 34a4940ab956be2)
in conjunction with this. However, this change is correct regardless
of any gcc bug.

Reviewed-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200614152544.136228-1-avi@scylladb.com>