tests: perf: fair_queue_perf: prepare for when_all_succeed() returning a tuple
when_all_succeed() will soon return a future<tuple> instead of a variadic
future. Prepare for that by discarding the result of a when_all_succeed()
that currently returns a future<> (and is consumed as a future<>). The new
code works with both the old and the new when_all_succeed().