sstring: Generalize make_sstring a bit
With this change it is possible to use make_sstring with any type that
can be used to construct a std::string_view. In particular, it is now
possible to use make_sstring with std::string_view and "const char*".

This makes it easier to efficiently concatenate strings, regardless of
their representation.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200616220015.240570-1-espindola@scylladb.com>