future: Move the function in then/then_wrapped only once
Before this patch the function object passed to then is moved when
creating the lambda:

  schedule(fut.get_promise(), [func = std::forward<Func>(func)] ...

And when the lambda is moved into the continuation.

By having a continuation explicitly contain both the original function
object and a wrapper, we can move the original function object only
once.

This is a 0.7% reduction in the scylla text size. From 41492660 to
41799156.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200617184233.495941-1-espindola@scylladb.com>