future: Use a new seastar::nested_exception in finally
The std::nested_exception is designed to be inherited from. The
derived type holds the outer exception and the inner exception is
captured by the std::nested_exception constructor which calls
std::current_exception().

That is not a good match for future::finally. We don't have the
derived type anymore, just a std::exception_ptr.

This patch introduces a seastar::nested_exception that just holds two
std::exception_ptr and uses that in finally.

With this finally actually saves both exceptions.

Fixes #760

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200618005650.936695-1-espindola@scylladb.com>