io-queue: Dont leak desc
When queueing a desc+func into fair-queue by the time _fq.queue()
is called the desc's raw pointer is already captured on lambda
and the unique pointer is released.

The queue() can throw allocating memory, thus the desc memory
will remain not released. The unique_ptr was supposed to handle
this, but the .release happens too early.

Fix by keeping the pointer on lambda and releasing it inside.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20200618085935.3823-2-xemul@scylladb.com>