smp: specify submit_to as noexcept
Now that `this_shard_id` is noexcept, `submit_to(unsigned, smp_submit_to_options, Func&&)`
handles all exceptions and calls only noexcept functions hence
it can be specified noexcept.

It is called from `submit_to(unsigned, Func&&)` that makes
a `default_smp_service_group()` that is noexcept too,
and forwards the Func&&, so it can be specifies as noexcept as well.

As Avi Kivity <avi@scylladb.com> observed:
Note that while the underlying implementation of submit_to() is
noexcept, it does allocate and will terminate if the allocation fails.
This is not an allocation failure we can convert to an exceptional
future, since the application may rely on submit_to() in error recovery
paths (for example. to unlock a lock on another shard).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>