reactor: make rename_priority_class noexccept
Inherit from futurize_invoke.

The lambda passed to it is made nothrow move constructible
(modulo https://gcc.gnu.org/bugzilla/show_bug.cgi?id=95368).
in preparation to requiring that for making invoke_on_all
noexcept.  This "fixed" a hidden use-after-move that was masked
by the fact the new_name was captured as const so the std::move
to io_queue::rename_one_priority_class was a no-op.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>