future: Improve asserts in promise_base::clear
Since a9f0953e8ae96d7ec8a2429bf3661ffa07d3930f it is possible to have
both the _future and _task pointers set, so we should have two
independent if statements.

The assert

    assert(_state->available() || !_task);

Was also a bit too strict. It almost all cases, it is the "!_task"
part that was making it pass since only with thread_wake_task that we
get both a _future and a _task. With thread_wake_task one would have
to do some gymnastics to destroy the promise before the thread was
switched back in and the state was available.

That is about to become easier as the following patches in these
series make us treat coroutines a bit more like seastar::thread.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>