io_queue: Fix discrepancy of queuer reqs vs their number
When counting the requests queued in fair-queue the counters on
io_queue and priority_class_data are incremented first, then the
fair_queue::queue is called. The counters are decremented in the
queue()'s callback and since the former may throw, the counters
will remain incremented.

This is not a problem for io_queue's one, it's not used by anyone,
but the priority_class_data's one is reported into stats and is
thus used.

Fix by incrementing the counters after queue() call.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>