fair_queue: Fix _requests_executing accounting
The counter is inc-only in current code, but since nobody uses it, it
comes unnoticed. Shared fair queue will need it (in draft it needs) so
keep it correct. For this kill the batch processing, as it saves nothing
but keeps the io_queue class bigger, and notify the fq about completed
request right at once.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>