fair_queue: Shuffle dispatch/charge codeflow
Current dispatching logic is

    while (have_resources()) {
        req = peek_next_request();
        dequeue_request(req);
        account_and_dispatch(req);
    }

In the next patches the "heaviness" of the next request will be needed
by the have_resources() callback, so change the codeflow to look like

    while (true) {
        request = peek_next_request();
        if (!check_and_account_resources(req)) {
            break;
        }
        dequeue_request(req);
        dispatch(req);
    }

NB: Actually the need to know the size of the request when checking for
available capacity is needed not only for IO groups. Current code can
over-dispatch if the request is "large" enough, so that

  executing_capacity < current_capacity < executing_capacity + request.size

but that's to be resolved separately, the patched behavior is the same.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>