tests: Fix perf fair queue stuck
The local-queue test measures the throughput of fair-queue, to do
it it spawns two fibers:

1. queues requests into fair queue, when dispatched the request
increments the "executed" counter (this fiber is called invokers)

2. calls queue's dispatch loop inside do_until() waiting for the
"executed" counter becomes equal (not greater, but equal) to the goal
(this one is called collectors)

Also
 - the test is run on the very same instance multiple times,
 - the counter in question is zeroed in the 1st fiber and
 - both fibers are "merged" with when_all_succeed()

Now the problem. If the 2nd (or later) iteration somehow manages to run
the collector fiber first, the "executed" counter becomes equal to the goal
right at once, because it is not _yet_ zeroed by the invoker fiber. As an
intermediate result the queue is _not_ be dispatched at this iteration,
but is stuffed with requests once the invokers is run.

Respectively, next iteration's invoker fiber will queue more requests int
it and once the queue dispatching starts the requests from it overflow the
"executed" counter and the collector fiber never stops.

The fix is to zero the "executed" counter in the 2nd fiber.
tests: Fix perf fair queue stuck

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20200706163812.11051-1-xemul@scylladb.com>