future: Refactor a bit of duplicated code
This moves a bit of duplicated code to future_base::schedule. The
produced code changes are insignificant by themselves: 0.07% text size
reduction and 0.1% tps improvement (probably noise).

The main advantage is that both allocations of a continuation (with or
without a ready state) are in the same function. This will be helpful
for the patch series dropping the calls to need_preempt.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200708054952.119179-1-espindola@scylladb.com>