futures: Add a test for a broken promise in repeat
This was fixed by 386f575021e5327155f3f4a5b0d580c80c2bd45e. That patch
claims to have no functionality change, but that was under the
assumption that we already had all the required calls to
set_to_broken_promise. It turns out that some cases, like the one
tested by this patch, were missing a call to set_to_broken_promise and
would crash without that patch.

I am adding a test with repeat in addition to the one with
parallel_for_each because when back porting I hit a revision where the
parallel_for_each test passes, but because of a change to the
parallel_for_each implementation
(951effa90ed47040827d0da17c7ee4045036249d), the underlying problem is
still there.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200713214828.55338-1-espindola@scylladb.com>