reactor_backend: let the reactor know again if any work was done by aio backend
After this refactoring 8ca7854e34e99, the aio backend is no longer
returning correct information about whether any work was performed.
The reactor will only think the backend did any work if it retries
requests that failed with EAGAIN. The statement, that says work was
performed after a request is submitted through io_submit(), was lost
in the refactoring.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20200717202511.27632-1-raphaelsc@scylladb.com>