fstream: turn file_data_source_impl's memory corruption bugs into aborts
We know for a fact that file_data_source_impl can easily corrupt memory if
it is created with read ahead enabled and its user forgets about closing
it properly. The fibers running on behalf of read ahead happen in background
and thus could easily write to freed memory if close() wasn't called.
That's because all the buffers used by the reads in progress would be
potentially freed by the time it's written to.

We've fixed something similar for append_challenged_posix_file_impl() but
this is only used in the context of writes. make_file_impl() creates a
posix_file_impl() instead if file is opened read-only.
We could have added this check to posix_file_impl dtor, but this memory
corruption issue mostly comes from the read ahead impl and that belongs
to file_data_source_impl and the latter knows better when to abort if
there are reads in progress.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20200727175226.371587-1-raphaelsc@scylladb.com>