reactor: Pass a std::string_view to file operations
Having a noexcept function that takes a sstring is a bit error
prone. Calling

  func("bar").then(...)

Seems fine, but than actually throw in the sstring constructor throws.

This patch changes the signature to use std::string_view and makes the
existing copies in the implementation explicit.

The next patch will handle std::bad_alloc in the implementation.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>