future: Simplify future::schedule
With recent changes in in master, future::schedule is only called with
unavailable futures. Use that to simplify the code a bit.

This change unfortunately increases the scylla text by 0.44%, from
37410436 to 37573988 bytes.

As far as I can tell the increase is because we were hitting some
limits in the gcc inlining heuristics, and this simplifications causes
more overall inlining.

If I first mark schedule noinline, the text size grows to 38674852,
and this change brings it down to 37934532.

perf_simple_query report a 0.13% tps improvement over 300 randomized
runs (form 160024 to 160241), but that can also just be noise.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200801183449.4234-1-espindola@scylladb.com>