reactor: Create groups by index and attach queues to them
In the previous patch for each io_queue reactor allocated new io_group.
The goal of grouping is to have N:1 relation, so allocate groups on
demand based on group ID assigned by resources.cc.

For now resources.cc ties all shards to a single node-wide group, but
this will also change in next patches. This patch goal is to throw the
queue-to-group assignment logic into the game.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>