test: Increase capacity of fair-queue unit test case
The test sets queue limit to 1 and feeds it with requests of weights
1 and 2. Current dispatcher allows one request to over-dispatch the queue,
so it lets one request with weight 2 go through. Technically this is not
correct, the queue capacity should allow for the biggest request to pass
by its own. Also this becomes critical corner-case for shared io queues,
where the over-dispatch is not welcome.

tests: unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20200806091831.13979-1-xemul@scylladb.com>