memory: fix small aligned free memory corruption
Over-aligned free is invoked by the compiler when calling
operator delete on an over-aligned object, with the same
parameters as the over-aligned allocation (using operator new
and operator delete overloads with an std::align_val_t parameter).

However, the implementations are not symmetric: aligned allocation
adjusts the size to the next power of two to compenstate for the
fact that the small pool allocator only guarantees 16 byte alignment
for arbitrary sizes, but does guarantee full alignment for
power-of-two sizes. The aligned free function, however, did not
do this adjustment, so it used a mismatching size when selecting
the small_pool into which to free. This caused corruption in the
internal data structures, easily observed as small_pool::_pages_in_use
going negative.

Fix by making aligned_free() perform the same adjustment as
aligned_alloc().
Message-Id: <20200809090540.2613939-1-avi@scylladb.com>