fair_group: Introduce rovers
The rover is monotonically increasing ticket. Overflow is possible and not
tried to be avoided. Instead, for two rovers it's possible to get if either
of them is ahead of the other, provided they didn't ran away from each other
too far.

This is rovers who suffer(ed) from the overflow problem mention in the patch
that switched tickets on the 512-block accounting.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>